fixing fw_update

+++ bootloader/src/bootloader.c
@@ -217,7 +217,7 @@ void load_firmware(uint32_t interface, uint32_t size){
         for(j = 0; j < frame_size; j++){
             firmware_buffer[j + pos] = page_buffer[j];
         }
-        pos += FLASH_PAGE_SIZE;
+        pos += (uint32_t)FLASH_PAGE_SIZE;
         remaining -= frame_size;
         j = 0;
     }
